Support new ECR credential ID format
The Amazon ECR plugin now stores the region ID in the credentials ID
(ecr:region:id). Since this is now available there, the old UI element
used to specify the ECR region has been removed and it defaults to
us-east-1 if not specified in the credentials (to support the old format,
like the Amazon ECR plugin does). Plugin now detects ECR credentials
automatically instead of using the old UI element.